# swagger/v1/schema.yaml
...
openapi: 3.1.0
info:
  title: Reckoning API
  version: v1
  license:
    name: GNU General Public License v3.0
    url: https://github.com/mortik/reckoning/blob/main/LICENSE
paths:
...
  "/customers/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      required: true
    get:
      summary: Get Customer
      tags:
      - Customers
      operationId: customer
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Customer"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          "$ref": "#/components/responses/SchemaValidationError"
    patch:
      summary: Update Customer
      tags:
      - Customers
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CustomerInput"
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Customer"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
    delete:
      summary: Destroy Customer
      tags:
      - Customers
      operationId: destroyCustomer
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/invoices/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      required: true
    get:
      summary: Get Invoice
      tags:
      - Invoices
      operationId: invoice
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Invoice"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          "$ref": "#/components/responses/SchemaValidationError"
    patch:
      summary: Update Invoice
      tags:
      - Invoices
      operationId: updateInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/InvoiceInput"
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Invoice"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
    delete:
      summary: Destroy Invoice
      tags:
      - Invoices
      operationId: destroyInvoice
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Message"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/projects/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      required: true
    get:
      summary: Get Project
      tags:
      - Projects
      operationId: project
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Project"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          "$ref": "#/components/responses/SchemaValidationError"
    patch:
      summary: Update Project
      tags:
      - Projects
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProjectInput"
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Project"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
    delete:
      summary: Destroy Project
      tags:
      - Projects
      operationId: destroyProject
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Message"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
servers:
- url: "/api/v1"
  description: Current host
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type:
          - string
          - 'null'
        country:
          type:
          - string
          - 'null'
        email:
          type:
          - string
          - 'null'
        telefon:
          type:
          - string
          - 'null'
        fax:
          type:
          - string
          - 'null'
        website:
          type:
          - string
          - 'null'
        paymentDue:
          type:
          - integer
          - 'null'
        invoiceEmail:
          type:
          - string
          - 'null'
        invoiceEmailCc:
          type:
          - string
          - 'null'
        invoiceEmailBcc:
          type:
          - string
          - 'null'
        defaultFrom:
          type:
          - string
          - 'null'
        emailTemplate:
          type:
          - string
          - 'null'
        offerDisclaimer:
          type:
          - string
          - 'null'
        employmentDate:
          type:
          - string
          - 'null'
          format: date
        employmentEndDate:
          type:
          - string
          - 'null'
          format: date
        weeklyHours:
          type:
          - integer
          - 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - name
      - createdAt
      - updatedAt
      title: Customer
    CustomerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        country:
          type: string
        email:
          type: string
        telefon:
          type: string
        fax:
          type: string
        website:
          type: string
        paymentDue:
          type: integer
        invoiceEmail:
          type: string
        invoiceEmailCc:
          type: string
        invoiceEmailBcc:
          type: string
        defaultFrom:
          type: string
        emailTemplate:
          type: string
        offerDisclaimer:
          type: string
        employmentDate:
          type: string
          format: date
        employmentEndDate:
          type: string
          format: date
        weeklyHours:
          type: integer
      additionalProperties: false
      required:
      - name
      title: CustomerInput
...
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ref:
          type:
          - integer
          - 'null'
        refNumber:
          type:
          - string
          - 'null'
        title:
          type:
          - string
          - 'null'
        state:
          type:
          - string
          - 'null'
          enum:
          - created
          - charged
          - paid
          -
        date:
          type:
          - string
          - 'null'
          format: date
        deliveryDate:
          type:
          - string
          - 'null'
          format: date
        paymentDueDate:
          type:
          - string
          - 'null'
          format: date
        payDate:
          type:
          - string
          - 'null'
          format: date-time
        value:
          type:
          - string
          - 'null'
        vat:
          type:
          - string
          - 'null'
        customerId:
          type:
          - string
          - 'null'
          format: uuid
        customerName:
          type:
          - string
          - 'null'
        projectId:
          type:
          - string
          - 'null'
          format: uuid
        projectName:
          type:
          - string
          - 'null'
        editable:
          type: boolean
          description: False once the invoice has been charged.
        sendable:
          type: boolean
          description: The customer has an invoice email and template.
        positions:
          type: array
          items:
            "$ref": "#/components/schemas/InvoicePosition"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - state
      - positions
      - editable
      - sendable
      - createdAt
      - updatedAt
      title: Invoice
    InvoiceInput:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        ref:
          type:
          - integer
          - 'null'
        date:
          type:
          - string
          - 'null'
          format: date
        delivery_date:
          type:
          - string
          - 'null'
          format: date
        payment_due_date:
          type:
          - string
          - 'null'
          format: date
        positions_attributes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              description:
                type:
                - string
                - 'null'
              hours:
                type:
                - string
                - number
                - 'null'
              rate:
                type:
                - string
                - number
                - 'null'
              value:
                type:
                - string
                - number
                - 'null'
              timer_ids:
                type: array
                items:
                  type: string
                  format: uuid
              _destroy:
                type: boolean
            additionalProperties: false
      additionalProperties: false
      title: InvoiceInput
    InvoicePosition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type:
          - string
          - 'null'
        hours:
          type:
          - string
          - 'null'
        rate:
          type:
          - string
          - 'null'
        value:
          type:
          - string
          - 'null'
        timerIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - timerIds
      - createdAt
      - updatedAt
      title: InvoicePosition
...
    Message:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
      required:
      - message
      title: Message
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        customerName:
          type:
          - string
          - 'null'
        label:
          type: string
        tasks:
          type: array
          items:
            "$ref": "#/components/schemas/ProjectTask"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        links:
          type: object
          properties:
            show:
              type: object
              properties:
                href:
                  type: string
              required:
              - href
          additionalProperties: false
      additionalProperties: false
      required:
      - id
      - name
      - tasks
      - createdAt
      - updatedAt
      title: Project
    ProjectInput:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        rate:
          type:
          - string
          - number
        budget:
          type:
          - string
          - number
        budget_hours:
          type:
          - string
          - number
        budget_on_dashboard:
          type: boolean
        invoice_addition:
          type:
          - string
          - 'null'
        tasks_attributes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              billable:
                type: boolean
              _destroy:
                type: boolean
            additionalProperties: false
      additionalProperties: false
      required:
      - name
      title: ProjectInput
    ProjectTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        label:
          type: string
        billable:
          type: boolean
        projectId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - name
      - projectId
      - createdAt
      - updatedAt
      title: ProjectTask
...
    StandardError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
      required:
      - message
      title: StandardError
...
    ValidationError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      additionalProperties: false
      required:
      - code
      - message
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /sessions. Used by non-browser clients.
    SessionCookie:
      type: apiKey
      in: cookie
      name: RECKONING_TEST
      description: Devise session cookie. Used by the web frontend.
  responses:
    SchemaValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
            required:
            - error
            - details
security: []